[REFACTOR] app 폴더 내부의 파일에 use client 제거
- app 폴더 내부의 파일에 use client 제거
- PostDeatail 컴포넌트 생성
- CommentInput 컴포넌트로 clear 로직 옮기기
- 게시글 작성 페이지에 use client 제거
- Unauthorized 페이지 분리
