+++ kakaobase/src/app/post/[postId]/page.tsx
@@ -1,44 +1,14 @@
-'use client';
 import Header from '@/widgets/header/Header';
-import Loading from '@/shared/ui/Loading';
-import MiddleBar from '@/features/feeds/ui/MiddleBar';
 import CommentInput from '@/features/feeds/comments/ui/CommentInput';
-import ListRouter from '@/features/feeds/ui/ListRouter';
-import PostCard from '@/features/feeds/ui/PostCard';
-import usePostDetail from '@/features/feeds/posts/hooks/usePostDetailHook';
-import { recommentFormStateStore } from '@/features/feeds/comments/stores/recommentFormStateStore';
-import { useEffect } from 'react';
+import PostDetail from '@/features/feeds/ui/PostDetail';
 
 export default function Page({ params }: { params: { postId: number } }) {
   const id = Number(params.postId);
-  const { data, isPending } = usePostDetail({ id });
-  const { clear } = recommentFormStateStore();
-
-  useEffect(() => {
-    return () => clear();
-  }, []);
-
-  if (isPending)
-    return (
-      <div className="h-screen">
-        <Loading />
-      </div>
-    );
-  if (!data) return <div className="h-screen">게시글을 찾을 수 없습니다.</div>;
 
   return (
     <div className="flex flex-col h-screen scroll-none">
       <Header label="게시글 상세" />
-      <div
-        className="flex-grow flex flex-col h-screen w-full overflow-y-auto"
-        data-scroll-area
-      >
-        <div className="my-4 w-full">
-          <PostCard post={data} />
-        </div>
-        <MiddleBar />
-        <ListRouter />
-      </div>
+      <PostDetail postId={id} />
       <CommentInput />
     </div>
   );
