+++ kakaobase/src/app/post/new/page.tsx
@@ -1,4 +1,3 @@
-'use client';
 import Header from '@/widgets/header/Header';
 import PostEditor from '@/features/feeds/posts/ui/PostEditor';
