+++ kakaobase/src/features/feeds/ui/PostDetail.tsx
@@ -1,0 +1,31 @@
+'use client';
+import Loading from '@/shared/ui/Loading';
+import usePostDetail from '../posts/hooks/usePostDetailHook';
+import ListRouter from './ListRouter';
+import MiddleBar from './MiddleBar';
+import PostCard from './PostCard';
+
+export default function PostDetail({ postId }: { postId: number }) {
+  const { data, isPending } = usePostDetail({ id: postId });
+
+  if (!data) return <div className="h-screen">게시글을 찾을 수 없습니다.</div>;
+  if (isPending)
+    return (
+      <div className="h-screen">
+        <Loading />
+      </div>
+    );
+
+  return (
+    <div
+      className="flex-grow flex flex-col h-screen w-full overflow-y-auto"
+      data-scroll-area
+    >
+      <div className="my-4 w-full">
+        <PostCard post={data} />
+      </div>
+      <MiddleBar />
+      <ListRouter />
+    </div>
+  );
+}
